backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
supabase==2.3.0
PyJWT==2.8.0
redis==5.0.1
web3==6.11.3
stripe==7.8.0
numpy==1.24.3
python-jose[cryptography]==3.3.0